med_drf/serializers.py
from rest_framework import serializers

from med_drf.models import Article


class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = ('name', 'autor')
