pyproject.toml
[tool.poetry]
name = "med-website-drf"
version = "0.1.0"
description = ""
authors = ["Sergei Gorlanov <itserg90@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
django = "^5.0.6"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
